+++ index.js
@@ -159,6 +159,20 @@ module.exports = {
         }
       },
       rules: activeRulesConfig
+    },
+    'jsx-runtime': {
+      plugins: [
+        'react'
+      ],
+      parserOptions: {
+        ecmaFeatures: {
+          jsx: true
+        }
+      },
+      rules: {
+        'react/react-in-jsx-scope': 0,
+        'react/jsx-uses-react': 0
+      }
     }
   }
 };